refactor: move run_study to runner.py to eliminate circular dependency
folder.py was importing SimulationSession from session.py, which in turn
needed to import load_study from folder.py (worked around with a local
import). Moving run_study to a new runner.py breaks the cycle: folder.py
now only handles loading, session.py can import load_study at module level,
and runner.py owns the orchestration between the two.
